make storage.set work on intel HD5500, so tests pass ok

+++ src/lib/THClStorageGet.cl
@@ -1,4 +1,4 @@
-kernel void THClStorageGet(global float *res, global float *data, long index) {
+kernel void THClStorageGet(global float *res, global float *data, int index) {
   if(get_global_id(0) == 0) {
     res[0] = data[index];
   }
